# docs/api/onboarding.openapi.yaml
openapi: 3.1.0
info:
  title: The Home Land Onboarding API
  version: 0.1.0
  description: >-
    Contract for the Phase 1 onboarding walking skeleton. This contract is
    implementation-ready but remains subject to prototype and threat-model review.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /auth/registrations:
    post:
      tags: [Authentication]
      operationId: registerUser
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '202':
          description: Registration accepted without unsafe account disclosure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationAccepted'
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/email-verifications:
    post:
      tags: [Authentication]
      operationId: verifyEmail
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified and browser session established.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingSnapshot'
        '400': { $ref: '#/components/responses/InvalidToken' }
        '410': { $ref: '#/components/responses/ExpiredToken' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /organizations:
    post:
      tags: [Organizations]
      operationId: createOrganization
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization, owner membership, and onboarding progress created atomically.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationCreated'
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-thl_session
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 128 }
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 32, maxLength: 256 }
...
  schemas:
    Uuid:
      type: string
      format: uuid
    RegisterUserRequest:
      type: object
      additionalProperties: false
      required: [fullName, email, password, acceptedTermsVersion]
      properties:
        fullName: { type: string, minLength: 2, maxLength: 160 }
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 128 }
        acceptedTermsVersion: { type: string, minLength: 1, maxLength: 40 }
    RegistrationAccepted:
      type: object
      additionalProperties: false
      required: [status, nextAction]
      properties:
        status: { const: ACCEPTED }
        nextAction: { const: CHECK_EMAIL }
    VerifyEmailRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: { type: string, minLength: 32, maxLength: 512 }
    CreateOrganizationRequest:
      type: object
      additionalProperties: false
      required: [legalName, displayName, organizationType, primaryStateCode, approximateUnitRange]
      properties:
        legalName: { type: string, minLength: 2, maxLength: 200 }
        displayName: { type: string, minLength: 2, maxLength: 120 }
        organizationType:
          enum: [INDIVIDUAL_LANDLORD, PROPERTY_MANAGEMENT_COMPANY, REAL_ESTATE_AGENCY, INVESTMENT_GROUP]
        primaryStateCode: { type: string, pattern: '^[A-Z]{2}$' }
        approximateUnitRange:
          enum: [ONE_TO_NINE, TEN_TO_NINETY_NINE, ONE_HUNDRED_TO_FIVE_HUNDRED, FIVE_HUNDRED_TO_FIVE_THOUSAND, OVER_FIVE_THOUSAND]
...
    OnboardingSnapshot:
      type: object
      additionalProperties: false
      required: [state, nextAction, version]
      properties:
        organizationId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        state:
          enum: [REGISTERED, EMAIL_VERIFICATION_PENDING, EMAIL_VERIFIED, ORGANIZATION_CREATED, WORKSPACE_CONFIGURED, MFA_REQUIRED, PORTFOLIO_REQUIRED, READY_FOR_REVIEW, ACTIVE, ABANDONED, EXPIRED, SUSPENDED]
        nextAction: { type: string, maxLength: 80 }
        version: { type: integer, minimum: 1 }
    OrganizationCreated:
      type: object
      additionalProperties: false
      required: [organization, membershipRole, onboarding]
      properties:
        organization: { $ref: '#/components/schemas/OrganizationSummary' }
        membershipRole: { const: OWNER }
        onboarding: { $ref: '#/components/schemas/OnboardingSnapshot' }
    OrganizationSummary:
      type: object
      additionalProperties: false
      required: [id, displayName, slug, status, version]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        displayName: { type: string }
        slug: { type: [string, 'null'] }
        status: { enum: [ONBOARDING, ACTIVE, SUSPENDED, ARCHIVED] }
        version: { type: integer, minimum: 1 }
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: '^[A-Z0-9_]+$' }
        detail: { type: string }
        correlationId: { type: string }
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, code]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
  responses:
    Unauthenticated:
      description: Authentication required.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated actor is not authorized in this organization context.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ValidationError:
      description: Request failed semantic validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Current state conflicts with the requested command.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    InvalidToken:
      description: Verification token is invalid or already consumed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ExpiredToken:
      description: Verification token has expired.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    RateLimited:
      description: Request rate limit exceeded.
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }